Added SQLAlchemy database models
These are taken directly from the design in storyboard and translated to
sqlalchemy. Alembic is used and set up to do database migrations, and
the initial migration is created. openstack.common.db from
oslo.incubator and oslo.config are also set up.

The result of this is that there is a stories-db-manage command which
should be used to manage database migrations.